Moved AxisLine to common module.
